refactor(runtimed): rename native module to _internals, hide internal types
Rename the PyO3 native module from runtimed.runtimed to
runtimed._internals. The underscore signals this is not a supported
API surface.

NativeAsyncClient and AsyncSession are no longer re-exported from
the runtimed package — they're implementation details of Client and
Notebook. Users who need them can still reach them via
runtimed._internals, but the leading underscore makes the contract
clear.

Return-only data types (ExecutionResult, Output, Cell, RuntimeState,
etc.) remain importable from runtimed for type annotations.
