+++ python/runtimed/src/runtimed/__init__.py
@@ -7,17 +7,12 @@ from runtimed._cell import CellCollection, CellHandle
 # Primary API
 from runtimed._client import Client
 from runtimed._execution import Execution
-from runtimed._notebook import Notebook
-from runtimed._notebook_info import NotebookInfo
-from runtimed._presence import Presence
 
-# Data types (from native bindings)
-# These are importable but not in __all__ — they are return-only types
-# with no Python constructors. Users encounter them as return values
-# (e.g. cell.run() → ExecutionResult, notebook.runtime → RuntimeState)
-# but cannot instantiate them directly.
-from runtimed.runtimed import (  # noqa: F401
-    AsyncSession,
+# Return-only data types (from native bindings)
+# Importable for type annotations but not directly constructable.
+# Users encounter them as return values from the public API
+# (e.g. cell.run() → ExecutionResult, notebook.runtime → RuntimeState).
+from runtimed._internals import (  # noqa: F401
     Cell,
     CompletionItem,
     CompletionResult,
@@ -26,7 +21,6 @@ from runtimed.runtimed import (  # noqa: F401
     ExecutionResult,
     HistoryEntry,
     KernelState,
-    NativeAsyncClient,
     NotebookConnectionInfo,
     Output,
     PyQueueEntry,
@@ -39,6 +33,9 @@ from runtimed.runtimed import (  # noqa: F401
     show_notebook_app_for_channel,
     socket_path_for_channel,
 )
+from runtimed._notebook import Notebook
+from runtimed._notebook_info import NotebookInfo
+from runtimed._presence import Presence
 
 __all__ = [
     # Primary API — constructable entry points
