+++ python/runtimed/src/runtimed/__init__.pyi
@@ -3,21 +3,21 @@
 from runtimed._cell import CellCollection as CellCollection
 from runtimed._cell import CellHandle as CellHandle
 from runtimed._client import Client as Client
+from runtimed._internals import Cell as Cell
+from runtimed._internals import EnvState as EnvState
+from runtimed._internals import ExecutionEvent as ExecutionEvent
+from runtimed._internals import ExecutionResult as ExecutionResult
+from runtimed._internals import KernelState as KernelState
+from runtimed._internals import Output as Output
+from runtimed._internals import RuntimedError as RuntimedError
+from runtimed._internals import RuntimeState as RuntimeState
+from runtimed._internals import SyncEnvironmentResult as SyncEnvironmentResult
+from runtimed._internals import default_socket_path as default_socket_path
+from runtimed._internals import show_notebook_app as show_notebook_app
+from runtimed._internals import show_notebook_app_for_channel as show_notebook_app_for_channel
+from runtimed._internals import socket_path_for_channel as socket_path_for_channel
 from runtimed._notebook import Notebook as Notebook
 from runtimed._notebook_info import NotebookInfo as NotebookInfo
 from runtimed._presence import Presence as Presence
-from runtimed.runtimed import Cell as Cell
-from runtimed.runtimed import EnvState as EnvState
-from runtimed.runtimed import ExecutionEvent as ExecutionEvent
-from runtimed.runtimed import ExecutionResult as ExecutionResult
-from runtimed.runtimed import KernelState as KernelState
-from runtimed.runtimed import Output as Output
-from runtimed.runtimed import RuntimedError as RuntimedError
-from runtimed.runtimed import RuntimeState as RuntimeState
-from runtimed.runtimed import SyncEnvironmentResult as SyncEnvironmentResult
-from runtimed.runtimed import default_socket_path as default_socket_path
-from runtimed.runtimed import show_notebook_app as show_notebook_app
-from runtimed.runtimed import show_notebook_app_for_channel as show_notebook_app_for_channel
-from runtimed.runtimed import socket_path_for_channel as socket_path_for_channel
 
 __version__: str
